xztl: zone metadata module

+++ src/xapp-core.c
@@ -14,10 +14,10 @@ void xapp_print_mcmd (struct xapp_io_mcmd *cmd)
     printf ("naddr  : %d\n", cmd->naddr);
     printf ("status : %d\n", cmd->status);
     printf ("nlba0  : %lu\n", cmd->nlba[0]);
-    printf ("addr[0]: (%d/%d/%d/%d)\n", cmd->addr[0].g.grp,
-					    cmd->addr[0].g.punit,
-			                    cmd->addr[0].g.zone,
-				            cmd->addr[0].g.sect);
+    printf ("addr[0]: (%d/%d/%d/%lu)\n", cmd->addr[0].g.grp,
+				         cmd->addr[0].g.punit,
+			                 cmd->addr[0].g.zone,
+			      (uint64_t) cmd->addr[0].g.sect);
     printf ("prp0   : 0x%lx\n", cmd->prp[0]);
     printf ("callba : %s\n", (cmd->callback) ? "OK" : "NULL");
     printf ("async_c: %p\n", (void *) cmd->async_ctx);
